Consolidated general type-validation code into validateObject.ts to leave kacheryTypes.ts for more specific types.
