+++ src/figurl/viewInterface/kacheryTypes.ts
@@ -1,188 +1,7 @@
 import assert from 'assert';
-import crypto from 'crypto'
+import crypto from 'crypto';
+import validateObject, { isArrayOf, isBoolean, isJSONObject, isNumber, isObject, isObjectOf, isString, JSONObject, JSONValue, optional } from './validateObject';
 
-export type JSONPrimitive = string | number | boolean | null;
-export type JSONValue = JSONPrimitive | JSONObject | JSONArray;
-export type JSONObject = { [member: string]: JSONValue };
-export interface JSONArray extends Array<JSONValue> {}
-export const isJSONObject = (x: any): x is JSONObject => {
-    if (!isObject(x)) return false
-    return isJSONSerializable(x)
-}
-export const isJSONValue = (x: any): x is JSONValue => {
-    return isJSONSerializable(x)
-}
-export const tryParseJsonObject = (x: string): JSONObject | null => {
-    let a: any
-    try {
-        a = JSON.parse(x)
-    }
-    catch {
-        return null
-    }
-    if (!isJSONObject(a)) return null
-    return a;
-}
-export const isJSONSerializable = (obj: any): boolean => {
-    if (typeof(obj) === 'string') return true
-    if (typeof(obj) === 'number') return true
-    if (!isObject(obj)) return false
-    const isPlainObject = (a: Object) => {
-        return Object.prototype.toString.call(a) === '[object Object]';
-    };
-    const isPlain = (a: any) => {
-      return (a === null) || (typeof a === 'undefined' || typeof a === 'string' || typeof a === 'boolean' || typeof a === 'number' || Array.isArray(a) || isPlainObject(a));
-    }
-    if (!isPlain(obj)) {
-      return false;
-    }
-    for (let property in obj) {
-      if (obj.hasOwnProperty(property)) {
-        if (!isPlain(obj[property])) {
-          return false;
-        }
-        if (obj[property] !== null) {
-            if (typeof obj[property] === "object") {
-                if (!isJSONSerializable(obj[property])) {
-                    return false;
-                }
-            }
-        }
-      }
-    }
-    return true;
-}
-
-// object
-export const isObject = (x: any): x is Object => {
-    return ((x !== null) && (typeof x === 'object'));
-}
-
-// string
-export const isString = (x: any): x is string => {
-    return ((x !== null) && (typeof x === 'string'));
-}
-
-// function
-export const isFunction = (x: any): x is Function => {
-    return ((x !== null) && (typeof x === 'function'));
-}
-
-// number
-export const isNumber = (x: any): x is number => {
-    return ((x !== null) && (typeof x === 'number'));
-}
-
-// null
-export const isNull = (x: any): x is null => {
-    return x === null;
-}
-
-// boolean
-export const isBoolean = (x: any): x is boolean => {
-    return ((x !== null) && (typeof x === 'boolean'));
-}
-
-// isOneOf
-export const isOneOf = (testFunctions: Function[]): ((x: any) => boolean) => {
-    return (x) => {
-        for (let tf of testFunctions) {
-            if (tf(x)) return true;
-        }
-        return false;
-    }
-}
-
-export const optional = (testFunctionOrSpec: Function | ValidateObjectSpec): ((x: any) => boolean) => {
-    if (isFunction(testFunctionOrSpec)) {
-        const testFunction: Function = testFunctionOrSpec
-        return (x) => {
-            return ((x === undefined) || (testFunction(x)));
-        }
-    }
-    else {
-        return (x) => {
-            const obj: ValidateObjectSpec = testFunctionOrSpec
-            return ((x === undefined) || (_validateObject(x, obj)))
-        }
-    }   
-}
-
-// isEqualTo
-export const isEqualTo = (value: any): ((x: any) => boolean) => {
-    return (x) => {
-        return x === value;
-    }
-}
-
-// isArrayOf
-export const isArrayOf = (testFunction: (x: any) => boolean): ((x: any) => boolean) => {
-    return (x) => {
-        if ((x !== null) && (Array.isArray(x))) {
-            for (let a of x) {
-                if (!testFunction(a)) return false;
-            }
-            return true;
-        }
-        else return false;
-    }
-}
-
-// isObjectOf
-export const isObjectOf = (keyTestFunction: (x: any) => boolean, valueTestFunction: (x: any) => boolean): ((x: any) => boolean) => {
-    return (x) => {
-        if (isObject(x)) {
-            for (let k in x) {
-                if (!keyTestFunction(k)) return false;
-                if (!valueTestFunction(x[k])) return false;
-            }
-            return true;
-        }
-        else return false;
-    }
-}
-
-export type ValidateObjectSpec = {[key: string]: ValidateObjectSpec | (Function & ((a: any) => any))}
-
-export const _validateObject = (x: any, spec: ValidateObjectSpec, opts?: {callback?: (x: string) => any, allowAdditionalFields?: boolean}): boolean => {
-    const o = opts || {}
-    if (!x) {
-        o.callback && o.callback('x is undefined/null.')
-        return false;
-    }
-    if (typeof(x) !== 'object') {
-        o.callback && o.callback('x is not an Object.')
-        return false;
-    }
-    for (let k in x) {
-        if (!(k in spec)) {
-            if (!o.allowAdditionalFields) {
-                o.callback && o.callback(`Key not in spec: ${k}`)
-                return false;
-            }
-        }
-    }
-    for (let k in spec) {
-        const specK = spec[k];
-        if (isFunction(specK)) {
-            if (!specK(x[k])) {
-                o.callback && o.callback(`Problem validating: ${k}`)
-                return false;
-            }
-        }
-        else {
-            if (!(k in x)) {
-                o.callback && o.callback(`Key not in x: ${k}`)
-                return false;
-            }
-            if (!_validateObject(x[k], specK as ValidateObjectSpec, {callback: o.callback})) {
-                o.callback && o.callback(`Value of key > ${k} < itself failed validation.`)
-                return false;
-            }
-        }
-    }
-    return true;
-}
 
 // objectToMap and mapToObject
 export const objectToMap = <KeyType extends String, ValueType>(obj: {[key: string]: any}) => {
@@ -300,7 +119,7 @@ export interface Address {
     url?: UrlString
 }
 export const isAddress = (x: any): x is Address => {
-    if (!_validateObject(x, {
+    if (!validateObject(x, {
         hostName: optional(isHostName),
         port: optional(isPort),
         url: optional(isUrlString)
@@ -369,7 +188,7 @@ export interface KeyPair {
     privateKey: PrivateKey
 }
 export const isKeyPair = (x: any) : x is KeyPair => {
-    return _validateObject(x, {
+    return validateObject(x, {
         publicKey: isPublicKey,
         privateKey: isPrivateKey
     });
@@ -589,7 +408,7 @@ export interface FileKey {
 }
 
 export const isFileKey = (x: any): x is FileKey => {
-    return _validateObject(x, {
+    return validateObject(x, {
         sha1: isSha1Hash,
         manifestSha1: optional(isSha1Hash),
         chunkOf: optional({
@@ -616,7 +435,7 @@ export interface FindLiveFeedResult {
     nodeId: NodeId
 }
 export const isFindLiveFeedResult = (x: any): x is FindLiveFeedResult => {
-    return _validateObject(x, {
+    return validateObject(x, {
         nodeId: isNodeId
     });
 }
@@ -628,7 +447,7 @@ export interface FindFileResult {
     fileSize: ByteCount
 }
 export const isFindFileResult = (x: any): x is FindFileResult => {
-    if (!_validateObject(x, {
+    if (!validateObject(x, {
         nodeId: isNodeId,
         fileKey: isFileKey,
         fileSize: isByteCount
@@ -717,7 +536,7 @@ export interface SignedSubfeedMessage {
     signature: Signature
 }
 export const isSignedSubfeedMessage = (x: any): x is SignedSubfeedMessage => {
-    if (! _validateObject(x, {
+    if (! validateObject(x, {
         body: {
             previousSignature: optional(isSignature),
             messageNumber: isNumber,
@@ -788,7 +607,7 @@ export interface SubfeedWatch {
     channelName: ChannelName
 }
 export const isSubfeedWatch = (x: any): x is SubfeedWatch => {
-    return _validateObject(x, {
+    return validateObject(x, {
         feedId: isFeedId,
         subfeedHash: isSubfeedHash,
         position: isSubfeedPosition,
@@ -888,7 +707,7 @@ export interface FileManifestChunk {
     sha1: Sha1Hash
 }
 export const isFileManifestChunk = (x: any): x is FileManifestChunk => {
-    return _validateObject(x, {
+    return validateObject(x, {
         start: isByteCount,
         end: isByteCount,
         sha1: isSha1Hash
@@ -901,7 +720,7 @@ export interface FileManifest {
     chunks: FileManifestChunk[]
 }
 export const isFileManifest = (x: any): x is FileManifest => {
-    return _validateObject(x, {
+    return validateObject(x, {
         size: isByteCount,
         sha1: isSha1Hash,
         chunks: isArrayOf(isFileManifestChunk)
@@ -973,7 +792,7 @@ export type UserConfig = {
 }
 
 export const isUserConfig = (x: any): x is UserConfig => {
-    return _validateObject(x, {
+    return validateObject(x, {
         admin: optional(isBoolean)
     }, {
         allowAdditionalFields: true
