+++ src/views/Composite/CompositeViewData.ts
@@ -1,6 +1,6 @@
 import { validateObject } from "figurl"
-import { isOneOf, isSha1Hash, optional, Sha1Hash } from "figurl/viewInterface/kacheryTypes"
-import { isArrayOf, isEqualTo, isNumber } from "figurl/viewInterface/validateObject"
+import { isSha1Hash, Sha1Hash } from "figurl/viewInterface/kacheryTypes"
+import { isArrayOf, isEqualTo, isNumber, isOneOf, optional } from "figurl/viewInterface/validateObject"
 import { isString } from "vega"
 
 type CVViewData = {
